+++ modules/augment/src/augmenter.cpp
@@ -19,12 +19,12 @@ void Augmenter::applyImages(InputArrayOfArrays imgs, OutputArrayOfArrays dstImgs
     dstImgs.create(imgs.cols(), 1, 0, -1, true);
     RNG rng;
 
-    for (size_t i = 0; i < imgs.cols(); i++)
+    for (int i = 0; i < imgs.cols(); i++)
     {
         Mat originalImg = imgs.getMat(i);
         Mat img = originalImg.clone();
 
-            for (size_t j = 0; j < transformations.size(); j++)
+            for (int j = 0; j < transformations.size(); j++)
             {
                 float prob = rng.uniform(0.f, 1.f);
 
@@ -54,14 +54,14 @@ void Augmenter::applyImagesWithMasks(InputArrayOfArrays imgs,
     dstMasks.create(masks.cols(), 1, 0, -1, true);
     RNG rng;
 
-    for (size_t i = 0; i < imgs.cols(); i++)
+    for (int i = 0; i < imgs.cols(); i++)
     {
         Mat originalImg = imgs.getMat(i);
         Mat originalMask = masks.getMat(i);
         Mat augmentedImg = originalImg.clone();
         Mat augmentedMask = originalMask.clone();
 
-        for (size_t j = 0; j < transformations.size(); j++)
+        for (int j = 0; j < transformations.size(); j++)
         {
             float prob = rng.uniform(0.f, 1.f);
 
@@ -95,14 +95,14 @@ void Augmenter::applyImagesWithPoints(InputArrayOfArrays imgs,
     dstPoints.create(points.cols(), 1, 0, -1, true);
     RNG rng;
 
-    for (size_t i = 0; i < imgs.cols(); i++)
+    for (int i = 0; i < imgs.cols(); i++)
     {
         Mat originalImg = imgs.getMat(i);
         Mat originalPoints = points.getMat(i);
         Mat augmentedImg = originalImg.clone();
         Mat augmentedPoints = originalPoints.clone();
 
-        for (size_t j = 0; j < transformations.size(); j++)
+        for (int j = 0; j < transformations.size(); j++)
         {
             float prob = rng.uniform(0.f, 1.f);
 
@@ -136,14 +136,14 @@ void Augmenter::applyImagesWithRectangles(InputArrayOfArrays imgs,
     dstRects.create(rects.cols(), 1, 0, -1, true);
     RNG rng;
 
-    for (size_t i = 0; i < imgs.cols(); i++)
+    for (int i = 0; i < imgs.cols(); i++)
     {
         Mat originalImg = imgs.getMat(i);
         Mat originalRects = rects.getMat(i);
         Mat augmentedImg = originalImg.clone();
         Mat augmentedRects = originalRects.clone();
 
-        for (size_t j = 0; j < transformations.size(); j++)
+        for (int j = 0; j < transformations.size(); j++)
         {
             float prob = rng.uniform(0.f, 1.f);
